chore: align the ruff rule set with the sibling ha-* integrations
#131 pinned this repo to `["E4", "E7", "E9", "F"]` to stop ruff 0.16's
wider defaults from failing CI on untouched code, and said widening the
set was a deliberate cleanup rather than something a release decides.
This is that cleanup.

The set is now the one kakao-map and wardrowbe use (E/F/W, I, UP, B,
SIM, RUF, PL, ANN, ASYNC, TID) at line-length 100, target py314 — the
Python CI actually runs. That surfaced 460 findings; all are resolved.

Mechanical, no behavior change:
- pyupgrade/isort/typing: Optional[X] -> X | None, Dict/List -> dict/list,
  timezone.utc -> UTC, sorted imports
- TID252: parent-relative imports (`from ..const import ...`) are now
  absolute `custom_components.xbloom....`, matching wardrowbe. NB ruff's
  own unsafe fix resolves the root to bare `xbloom` — wrong for a
  custom_components layout — so this was done by hand
- ANN: annotated 58 signatures, mostly `device_info -> DeviceInfo` on the
  entity platforms and brewing.py's `client: XBloomClient`
- E501: wrapped 23 lines

Real, if small:
- PLC0415: 11 of the 12 function-level imports were leftovers whose
  module was already imported at the top of the same file; hoisted. The
  two in `llm/__init__.py` stay lazy on purpose (the platform must not
  pull in `homeassistant.components.llm` from the setup path) and now say
  so with a noqa
- SIM102/SIM105, RUF046/RUF059, B905/RUF007 fixed
- RUF012: `_attr_options` is HA's entity-attribute convention, not a
  ClassVar; noqa'd with the reason

Waived with justification rather than refactored:
- ASYNC109 globally — `timeout=` is forwarded to bleak or read back by
  the ACK retry logic, and the test doubles must mirror those signatures
- PLR complexity caps per file — protocol demultiplexers, decision tables
  and cloud-API mirrors whose size comes from the surface they mirror

`allowed-confusables` lets the LLM tool descriptions keep their en dashes
and × signs; those are prose, not identifiers.

Verified against both ruff 0.15.18 and 0.16.0 (0.16 stabilises PLR0917,
which lands on the two functions already waived for PLR0913). 449 passed,
3 skipped; every module still imports cleanly and the llm platform is
still not loaded at import time.

+++ custom_components/xbloom/llm/pour.py
@@ -130,7 +130,7 @@ class XBloomPourTool(XBloomBaseTool):
 
         # Mirror the slider state so the corresponding number entities reflect
         # what was actually requested.
-        self.coordinator.temperature = int(round(temperature))
+        self.coordinator.temperature = round(temperature)
         self.coordinator.volume = volume
         if flow_rate is not None:
             self.coordinator.flow_rate = float(flow_rate)
